[Incompatible] Rename module, internal storage changes
Rename module from Tree::Object to Tree::Object::Hash, because we'll also
provide other implementations like Tree::Object::Array, etc.

The main module Tree::Object now becomes empty and contains just a link to all
the available implementations.

Tree::Object::Hash: parent and children are now stored in '_parent' and
'_children' keys respectively instead of 'parent' and 'children', to denote that
they are internal and to reduce chance of clash.

Tree::Object::Hash: to set children, an arrayref must now be passed to
children() instead of list. This is because passing a list to set an arrayref
attribute is incompatible with accessor generators.

Tree::Object::Hash: children() now returns arrayref instead of list.